test/mochitest/test_steps_flashing.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1235459
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container"></div>
<pre id="test">
  <script type="application/javascript">

  "use strict";

  var allSteps = [ "downloading", "extracting", "fetching", "creating", "flashing"];

  function element_has_class(stepName, className) {
    let e = c.querySelector("." + stepName);
    ok(e, "has element", e);
    ok(e.classList.contains(className), "element has class", className);
  }

  function element_hasnot_class(stepName, classesName) {
    let e = c.querySelector("." + stepName);
    ok(e, "has element", e);

    ok(classesName.length > 0, "checking with classes");
    classesName.forEach(className => {
      ok(!e.classList.contains(className), "element has no class", className);
    });
  }

  function assert_step_inprogress(stepName) {
    element_has_class(stepName, "inprogress");
  }

  function assert_step_done(stepName) {
    element_has_class(stepName, "done");
  }

  function assert_step_fail(stepName) {
    element_has_class(stepName, "fail");
  }

  function assert_step_none(stepName) {
    element_hasnot_class(stepName, ["inprogress", "done", "fail"]);
  }

  function assert_dialog_visible(visible) {
    let computed = c.defaultView.getComputedStyle(c.getElementById("progressDialog"));

    if (visible) {
      is(computed.display, "block", "progress dialog is visible");
    } else {
      is(computed.display, "none", "progress dialog is not visible");
    }
  }

  function test_init() {
    return new Promise((resolve, reject) => {
      assert_dialog_visible(false);
      allSteps.forEach(s => assert_step_none(s));
      resolve();
    });
  }

  function test_reset() {
    return new Promise((resolve, reject) => {
      let dialog = c.getElementById("progressDialog")
      ok(dialog, "has dialog element");
      dialog.style.display = "none";

      allSteps.forEach(s => {
        let e = c.querySelector("." + s);
        ok(e, "has element", e);
        e.classList.remove("done");
        e.classList.remove("fail");
        e.classList.remove("inprogress");
      });

      resolve();
    });
  }

  function test_progress_step(stepName, info, subInfo, success = true) {
    return new Promise((resolve, reject) => {
      ok(true, "testing stepName: " + stepName + " as success:" + success);

      w.downloadProgress(stepName, info);
      if (!success) {
        w.downloadProgressFailure(info);
      }

      assert_dialog_visible(true);

      let currentStep = allSteps.indexOf(stepName);
      allSteps.forEach(s => {
        let isFinished = (allSteps.indexOf(s) < currentStep);
        let isFailed   = (allSteps.indexOf(s) === currentStep) && !success;
        let isRunning  = (allSteps.indexOf(s) === currentStep) && success;
        let isPending  = (allSteps.indexOf(s) > currentStep);

        if (isFinished) {
          assert_step_done(s);
        } else if (isRunning) {
          assert_step_inprogress(s);
        } else if (isFailed) {
          assert_step_fail(s);
        } else if (isPending) {
          assert_step_none(s);
        } else {
          ok(false, "unexpected state!!");
        }
      });

      if (info) {
        let e = c.getElementById("additionalProgress");
        is(e.textContent, info, "download info is here");
      }

      if (subInfo) {
        let e = c.getElementById("subAdditionalProgress");
        is(e.textContent, subInfo, "download sub info is here");
      }

      resolve();
    });
  }

  function test_downloading() {
    return test_progress_step("downloading", "Downloading archive");
  }

  function test_downloading_fail() {
    return test_progress_step("downloading", "Downloading archive FAILED", "", false);
  }

  function test_extracting() {
    return test_progress_step("extracting", "Extracting archive");
  }

  function test_extracting_fail() {
    return test_progress_step("extracting", "Extracting archive FAILED", "", false);
  }

  function test_fetching() {
    return test_progress_step("fetching", "Fetching blobs");
  }

  function test_fetching_fail() {
    return test_progress_step("fetching", "Fetching blobs FAILED", "", false);
  }

  function test_creating() {
    return test_progress_step("creating", "Creating images");
  }

  function test_creating_fail() {
    return test_progress_step("creating", "Creating images FAILED", "", false);
  }

  function test_flashing() {
    return test_progress_step("flashing", "Flashing device");
  }

  function test_flashing_fail() {
    return test_progress_step("flashing", "Flashing device FAILED", "", false);
  }

  function runTest() {
    populateAboutB2GInstaller()
      .then(() => test_init())
      .then(() => test_downloading())
      .then(() => test_extracting())
      .then(() => test_fetching())
      .then(() => test_creating())
      .then(() => test_flashing())
      /* Restart the process and check for some errors handling */
      .then(() => test_reset())
      /* First: error during download */
      .then(() => test_init())
      .then(() => test_downloading_fail())
      .then(() => test_reset())
      /* Second: error during extraction */
      .then(() => test_init())
      .then(() => test_downloading())
      .then(() => test_extracting_fail())
      .then(() => test_reset())
      /* Third: error during blob fetching */
      .then(() => test_init())
      .then(() => test_downloading())
      .then(() => test_extracting())
      .then(() => test_fetching_fail())
      .then(() => test_reset())
      /* Fourth: error during image creation */
      .then(() => test_init())
      .then(() => test_downloading())
      .then(() => test_extracting())
      .then(() => test_fetching())
      .then(() => test_creating_fail())
      .then(() => test_reset())
      /* Fifth: error during device flashing */
      .then(() => test_init())
      .then(() => test_downloading())
      .then(() => test_extracting())
      .then(() => test_fetching())
      .then(() => test_creating())
      .then(() => test_flashing_fail())
      .then(() => test_reset())
      .then(() => SimpleTest.finish());
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
